+++ packages/vercel-sandbox/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Minor Changes
 
 - Support a new method: Sandbox.getOrCreate()
+- Support pagination (CLI and SDK) when listing sandboxes, snapshots, sessions
 
 ## 2.0.0-beta.15